+++ packages/xrpc/src/xrpc-client.ts
@@ -1,11 +1,13 @@
 import { LexiconDoc, Lexicons, ValidationError } from '@atproto/lexicon'
 import {
   FetchHandler,
+  FetchHandlerObject,
   FetchHandlerOptions,
   buildFetchHandler,
 } from './fetch-handler'
 import {
   CallOptions,
+  Gettable,
   QueryParams,
   ResponseType,
   XRPCError,
@@ -14,6 +16,7 @@ import {
   httpResponseCodeToEnum,
 } from './types'
 import {
+  combineHeaders,
   constructMethodCallHeaders,
   constructMethodCallUrl,
   encodeMethodCallBody,
@@ -24,10 +27,11 @@ import {
 
 export class XrpcClient {
   readonly fetchHandler: FetchHandler
+  readonly headers = new Map<string, Gettable<null | string>>()
   readonly lex: Lexicons
 
   constructor(
-    fetchHandlerOpts: FetchHandler | FetchHandlerOptions,
+    fetchHandlerOpts: FetchHandler | FetchHandlerObject | FetchHandlerOptions,
     // "Lexicons" is redundant here (because that class implements
     // "Iterable<LexiconDoc>") but we keep it for explicitness:
     lex: Lexicons | Iterable<LexiconDoc>,
@@ -35,6 +39,18 @@ export class XrpcClient {
     this.fetchHandler = buildFetchHandler(fetchHandlerOpts)
 
     this.lex = lex instanceof Lexicons ? lex : new Lexicons(lex)
+  }
+
+  setHeader(key: string, value: Gettable<null | string>): void {
+    this.headers.set(key.toLowerCase(), value)
+  }
+
+  unsetHeader(key: string): void {
+    this.headers.delete(key.toLowerCase())
+  }
+
+  clearHeaders(): void {
+    this.headers.clear()
   }
 
   async call(
@@ -65,7 +81,7 @@ export class XrpcClient {
     // anywhere in docs or types. See whatwg/fetch#1438, nodejs/node#46221.
     const init: RequestInit & { duplex: 'half' } = {
       method: reqMethod,
-      headers: reqHeaders,
+      headers: combineHeaders(reqHeaders, this.headers),
       body: reqBody,
       duplex: 'half',
       signal: opts?.signal,